bikeways/serializers.py
from rest_framework import serializers

from .models import Bikeway, BikewayCategory


class BikewaySerializer(serializers.ModelSerializer):
    class Meta:
        model = Bikeway
        fields = ('id', 'name', 'location', 'condition', 'category',
                  'length')


class BikewayCategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = BikewayCategory
        fields = ('name', 'is_separated')
